Inserindo um novo Cliente com POST

+++ src/main/java/br/com/project/domain/Cliente.java
@@ -29,7 +29,7 @@ public class Cliente implements Serializable{
 	private String nome;
 	private String email;
 	private String cpfOuCnpj;
-	private Integer tipo;
+	private Long tipo;
 	
 	@OneToMany(mappedBy = "cliente")
 	private List<Endereco> enderecos = new ArrayList<>();
